feat: Application으로부터 사용자 처리 부분 controller로 분리
- Application은 의존성 관리만 하고, 사용자 요청을 처리하는 부분은 Controller에서 수행하도록 함
